move references from introduction (#96)
* move references from introduction

when moving ref from introduction, I forgot to copy them to the relevant section. This is fixed here + sorted in alphabetical order

* Update DACs_references.md

I added a section for the table, use the clean citing way and remove the link which should be redundant when the bib file will be updated along with the reference list.